# backend/openapi.yaml
openapi: 3.0.3
info:
  title: ShirazTyres API
  version: 1.0.0
  description: Emergency tyre callout platform — website, admin panel and mobile app.
paths:
...
  /api/v1/vehicles/{plate}/confirm-tyre:
    post:
      operationId: vehicles_confirm_tyre_create
      description: Staff-side vehicle browser plus the manual tyre-size override.
      parameters:
      - in: path
        name: plate
        schema:
          type: string
        required: true
      tags:
      - vehicles
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/TyreConfirm'
          application/x-www-form-urlencoded:
            schema:
              $ref: '#/components/schemas/TyreConfirm'
          multipart/form-data:
            schema:
              $ref: '#/components/schemas/TyreConfirm'
        required: true
      security:
      - staffJWT: []
      - driverJWT: []
      responses:
        '200':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Vehicle'
          description: ''
  /api/v1/vehicles/{plate}/refresh:
    post:
      operationId: vehicles_refresh_create
      description: Staff-side vehicle browser plus the manual tyre-size override.
      parameters:
      - in: path
        name: plate
        schema:
          type: string
        required: true
      tags:
      - vehicles
      security:
      - staffJWT: []
      - driverJWT: []
      responses:
        '200':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Vehicle'
          description: ''
components:
  schemas:
...
    TyreConfirm:
      type: object
      properties:
        tyre_size:
          type: string
          maxLength: 32
      required:
      - tyre_size
    TyreSourceEnum:
      enum:
      - api
      - driver
      - staff
      - unknown
      type: string
      description: |-
        * `api` - Tyre data API
        * `driver` - Confirmed by driver
        * `staff` - Entered by staff
        * `unknown` - Unknown
    Vehicle:
      type: object
      properties:
        id:
          type: integer
          readOnly: true
        plate:
          type: string
          maxLength: 16
        display_plate:
          type: string
          readOnly: true
        description:
          type: string
          readOnly: true
        make:
          type: string
          readOnly: true
        model:
          type: string
          readOnly: true
        colour:
          type: string
          readOnly: true
        fuel_type:
          type: string
          readOnly: true
        engine_capacity:
          type: integer
          readOnly: true
          nullable: true
          description: cc
        year_of_manufacture:
          type: integer
          readOnly: true
          nullable: true
        co2_emissions:
          type: integer
          readOnly: true
          nullable: true
        tax_status:
          type: string
          readOnly: true
        tax_due_date:
          type: string
          format: date
          readOnly: true
          nullable: true
        mot_status:
          type: string
          readOnly: true
        mot_expiry_date:
          type: string
          format: date
          readOnly: true
          nullable: true
        tyre_size_front:
          type: string
          readOnly: true
          description: e.g. 205/55R16
        tyre_size_rear:
          type: string
          readOnly: true
        tyre_load_index:
          type: string
          readOnly: true
        tyre_speed_rating:
          type: string
          readOnly: true
        tyre_pressure_front_psi:
          type: integer
          readOnly: true
          nullable: true
        tyre_pressure_rear_psi:
          type: integer
          readOnly: true
          nullable: true
        tyre_size_options:
          readOnly: true
          description: Alternative fitments when the trim is ambiguous.
        tyre_source:
          allOf:
          - $ref: '#/components/schemas/TyreSourceEnum'
          readOnly: true
        dvla_fetched_at:
          type: string
          format: date-time
          readOnly: true
          nullable: true
        tyre_fetched_at:
          type: string
          format: date-time
          readOnly: true
          nullable: true
        lookup_error:
          type: string
          readOnly: true
        updated_at:
          type: string
          format: date-time
          readOnly: true
      required:
      - co2_emissions
      - colour
      - description
      - display_plate
      - dvla_fetched_at
      - engine_capacity
      - fuel_type
      - id
      - lookup_error
      - make
      - model
      - mot_expiry_date
      - mot_status
      - plate
      - tax_due_date
      - tax_status
      - tyre_fetched_at
      - tyre_load_index
      - tyre_pressure_front_psi
      - tyre_pressure_rear_psi
      - tyre_size_front
      - tyre_size_options
      - tyre_size_rear
      - tyre_source
      - tyre_speed_rating
      - updated_at
      - year_of_manufacture
  securitySchemes:
    driverJWT:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: JWT issued for the driver audience.
    staffJWT:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: JWT issued for the staff audience.